+++ cmd/slack-status-changer/main.go
@@ -1,0 +1,33 @@
+package main
+
+func main() {
+	// slack := changer.Slack{}
+	// homedir, err := homedir.Dir()
+	// if err != nil {
+	// 	log.Fatal(err)
+	// }
+	// data, err := ioutil.ReadFile(homedir + "/.slack.yml")
+	// if err != nil {
+	// 	log.Fatal(err)
+	// }
+	// err = yaml.Unmarshal(data, &slack)
+	// if err != nil {
+	// 	log.Fatal(err)
+	// }
+	// client := &http.Client{
+	// 	Transport: &http.Transport{
+	// 		TLSClientConfig: &tls.Config{
+	// 			InsecureSkipVerify: true,
+	// 		},
+	// 	},
+	// }
+	// resp, err := client.Get(slack.Endpoint)
+	// if err != nil {
+	// 	log.Fatal(err)
+	// }
+	// if resp.StatusCode == 200 {
+	// 	slack.sendStatus("office")
+	// } else {
+	// 	slack.sendStatus("house")
+	// }
+}
